refactor: event-memer -> evevt-member 파일이름 변경
